requirements.txt
tensorflow
Pillow
keras
scipy
tensorflow_hub
streamlit
opencv-python